Change order of episode selection
An issue was brought to my attention wherein if an episode is not
available in the selected quality, the user would not know that it
existed at all.  This is made evident in the example of the PsychoPass
first season (which is only available in 720p).  This commit changes the
order that information is presented to the user, showing all the
shows followed by the qualities that these shows are available in.